Move logger initialization to very early in startup
