+++ engine/src/Core/Core.hpp
@@ -20,11 +20,11 @@
 #define KOKKO_PROFILE_FUNCTION()
 #endif
 
-#include "Debug/Log.hpp"
+#include "System/Log.hpp"
 
-#define KK_LOG_DEBUG(format, ...) ::Log::LogDebug(__FILE__, __LINE__, FMT_STRING(format), ## __VA_ARGS__)
-#define KK_LOG_INFO(format, ...) ::Log::Log(LogLevel::Info, FMT_STRING(format), ## __VA_ARGS__)
-#define KK_LOG_WARN(format, ...) ::Log::Log(LogLevel::Warning, FMT_STRING(format), ## __VA_ARGS__)
-#define KK_LOG_ERROR(format, ...) ::Log::Log(LogLevel::Error, FMT_STRING(format), ## __VA_ARGS__)
+#define KK_LOG_DEBUG(format, ...) kokko::Log::LogDebug(__FILE__, __LINE__, FMT_STRING(format), ## __VA_ARGS__)
+#define KK_LOG_INFO(format, ...) kokko::Log::Log(LogLevel::Info, FMT_STRING(format), ## __VA_ARGS__)
+#define KK_LOG_WARN(format, ...) kokko::Log::Log(LogLevel::Warning, FMT_STRING(format), ## __VA_ARGS__)
+#define KK_LOG_ERROR(format, ...) kokko::Log::Log(LogLevel::Error, FMT_STRING(format), ## __VA_ARGS__)
 
 #define KK_CACHE_LINE 64
